requirements.txt
# Core dependencies for AI and cloud integration
semantic-kernel>=1.0.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-network>=25.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web framework for API (optional)
fastapi>=0.100.0
uvicorn>=0.20.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Logging and utilities
structlog>=23.0.0
rich>=13.0.0
typer>=0.9.0